working on negative

+++ libpnt/ft_3_vector_len.c
@@ -14,5 +14,5 @@
 
 float			ft_3_vector_len(t_vector vec)
 {
-	return ((float)sqrt(vec[0] * vec[0] + vec[1] * vec[1] + vec[2] * vec[2]));
+	return (sqrtf(vec[0] * vec[0] + vec[1] * vec[1] + vec[2] * vec[2]));
 }
